+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Remove data_table from existing plan_record entity type definition #829](https://github.com/farmOS/farmOS/pull/829)
 - [Use farm_people entity_reference View for exposed Owner filter in farm_asset View #835](https://github.com/farmOS/farmOS/pull/835)
 - [Require asset name in bulk KML importer #836](https://github.com/farmOS/farmOS/pull/836)
+- Patch Drupal core to fix [Issue #3420372: Core Symfony Mailer throws error on transport shutdown](https://www.drupal.org/project/drupal/issues/3420372)
 
 ## [3.2.1] 2024-04-12