+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.2.2] 2024-05-17
 
 ### Added
 
@@ -776,7 +776,8 @@ moving forward.
     Drupal 7, which required a complete refactor of the codebase. By comparison,
     updating from Drupal 9 to 10 will simply involve updating deprecated code.
 
-[Unreleased]: https://github.com/farmOS/farmOS/compare/3.2.1...HEAD
+[Unreleased]: https://github.com/farmOS/farmOS/compare/3.2.2...HEAD
+[3.2.2]: https://github.com/farmOS/farmOS/releases/tag/3.2.2
 [3.2.1]: https://github.com/farmOS/farmOS/releases/tag/3.2.1
 [3.2.0]: https://github.com/farmOS/farmOS/releases/tag/3.2.0
 [3.1.2]: https://github.com/farmOS/farmOS/releases/tag/3.1.2